Fix clang-tidy warnings to pass CI tidy check
- Use emplace_back instead of push_back for aggregate types
- Add static_cast for implicit widening/narrowing conversions
- Use const reference for copy-constructed read-only variable

+++ src/functions/level_pivot_delete.cpp
@@ -102,7 +102,7 @@ SourceResultType LevelPivotDelete::GetData(ExecutionContext &context, DataChunk 
                                            OperatorSourceInput &input) const {
 	auto &gstate = sink_state->Cast<LevelPivotDeleteGlobalState>();
 	chunk.SetCardinality(1);
-	chunk.SetValue(0, 0, Value::BIGINT(gstate.delete_count));
+	chunk.SetValue(0, 0, Value::BIGINT(static_cast<int64_t>(gstate.delete_count)));
 	return SourceResultType::FINISHED;
 }
