+++ src/functions/level_pivot_update.cpp
@@ -110,7 +110,7 @@ SourceResultType LevelPivotUpdate::GetData(ExecutionContext &context, DataChunk 
                                            OperatorSourceInput &input) const {
 	auto &gstate = sink_state->Cast<LevelPivotUpdateGlobalState>();
 	chunk.SetCardinality(1);
-	chunk.SetValue(0, 0, Value::BIGINT(gstate.update_count));
+	chunk.SetValue(0, 0, Value::BIGINT(static_cast<int64_t>(gstate.update_count)));
 	return SourceResultType::FINISHED;
 }
